Initialisation de l'interface WPF, création des premiers liens entre les nouvelles Vues et les Vue-Modèles.

Related work items: #12

+++ EasySave_G8_UI/Models/Model_PRE.cs
@@ -13,9 +13,7 @@ namespace EasySave_G8_UI.Models
         public string Source { get; set; }
         public string Destination { get; set; }
         public bool Type { get; set; }
-        public Model_PRE()
-        {
-        }
+
         public Model_PRE(string Name, string Source, string Destination, bool Type)
         {
             this.Name = Name;
@@ -40,6 +38,12 @@ namespace EasySave_G8_UI.Models
                 string fileContent = File.ReadAllText(fileName); //Bring content of filename in filecontent
                 List<Model_PRE> ?values = new List<Model_PRE>(); //Create the list named values
                 values = JsonConvert.DeserializeObject<List<Model_PRE>>(fileContent); //Deserialialize the data in JSON form
+                if (values.Count() == 5 || values.Count() > 5) //Check if there is already 5 works
+                {
+                    View_Model ViewMODEL = new View_Model();
+                    ViewMODEL.MV_Work_5Works();
+                    return;
+                }
                 values?.Add(this); //Add object ModelAFT in the list values
                 string jsonString = Newtonsoft.Json.JsonConvert.SerializeObject(values, Newtonsoft.Json.Formatting.Indented); //Serialialize the data in JSON form
                 File.WriteAllText(fileName, jsonString); //Write json file
